refactor: move useMenu to its own folder.
