+++ docs/data/base/components/menu/UseMenu.js
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { useMenu, MenuUnstyledContext } from '@mui/base/MenuUnstyled';
+import useMenu from '@mui/base/useMenu';
+import { MenuUnstyledContext } from '@mui/base/MenuUnstyled';
 import useMenuItem from '@mui/base/useMenuItem';
 import { GlobalStyles } from '@mui/system';
 import clsx from 'clsx';
