+++ packages/mui-base/src/useMenu/index.ts
@@ -1,0 +1,5 @@
+export { default as useMenu } from './useMenu';
+export { default } from './useMenu';
+export * from './useMenu';
+
+export * from './useMenu.types';
